+++ CHANGELOG.md
@@ -9,6 +9,30 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 Now the whol `Webhook` instance is passed as first argument instead of shop order reference. Existing code can get
 the shop order id directly from `Webhook` instance.
 
+# [v1.0.10](https://github.com/sequra/integration-core/tree/v1.0.10)
+## Changed
+- Added order total calculation in cart items setter.
+
+# [v1.0.9](https://github.com/sequra/integration-core/tree/v1.0.9)
+## Changed
+- Added optional parameters to the updateStatus method of ShopOrderService interface.
+
+# [v1.0.8](https://github.com/sequra/integration-core/tree/v1.0.8)
+## Changed
+- Made methods of webhook handler protected.
+
+# [v1.0.7](https://github.com/sequra/integration-core/tree/v1.0.7)
+## Changed
+- Fixed a duplicated cart bug when transforming order update request.
+
+# [v1.0.6](https://github.com/sequra/integration-core/tree/v1.0.6)
+## Changed
+- Removed operator reference from merchant DTO.
+
+# [v1.0.5](https://github.com/sequra/integration-core/tree/v1.0.5)
+## Changed
+- Updated the endpoint URL for updating order carts in proxy.
+
 # [v1.0.4](https://github.com/sequra/integration-core/tree/v1.0.4)
 ## Changed
 - Updated the cart DTO to allow updating order items.